pyproject.toml
[project]
name = "bildungsurlaub-azureai"
version = "1.0"
description = "Bildungsurlaub-Kurs 'Azure AI' von Gollnick Data Solutions"
readme = "README.md"
requires-python = ">=3.12.9"
dependencies = [
    "azure-ai-documentintelligence>=1.0.2",
    "azure-ai-formrecognizer>=3.3.3",
    "azure-ai-inference>=1.0.0b9",
    "azure-ai-projects>=1.0.0",
    "azure-ai-textanalytics>=5.3.0",
    "azure-ai-translation-text>=1.0.1",
    "azure-ai-vision-imageanalysis>=1.0.0",
    "azure-cognitiveservices-speech>=1.46.0",
    "azure-cognitiveservices-vision-customvision>=3.1.1",
    "azure-identity>=1.25.0",
    "azure-keyvault>=4.2.0",
    "azure-search-documents>=11.6.0",
    "gradio>=5.46.0",
    "ipykernel>=6.30.1",
    "langchain>=1.0.3",
    "langchain-azure-storage>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.2",
    "langchain-openai>=1.0.1",
    "langchain-text-splitters>=1.0.0",
    "matplotlib>=3.10.6",
    "openai>=1.108.0",
    "pygame>=2.6.1",
    "pymupdf>=1.26.5",
    "python-dotenv>=1.1.1",
]
